Move weapons in weapons folder
