+++ src/weapons/__init__.py
@@ -1,0 +1,3 @@
+from src.weapons.projectile import Projectile
+from src.weapons.grenade import Grenade
+from src.weapons.rocket import Rocket
